Scope rename shortcut to be inside selected row
